chatterfix/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
google-cloud-firestore==2.20.2
python-dotenv==1.0.1
openai==1.8.0
pydantic==2.6.4
httpx==0.26.0
certifi==2025.4.26
